Fix isNodeModule check in Target entity
Updated the isNodeModule method to correctly return whether or not the module path exists. Previously, it was only returning true if the module path included 'node_modules' and the import source path, which is not always the case, especially with locally defined modules. Thus, this new approach covers more use cases.