Update pull request triggers in GitHub workflow
Changed the trigger for the GitHub Actions workflow in test.yaml. Instead of triggering on any push or pull request on the main branch, the workflow will now trigger when a pull request is opened, synchronized, or reopened. This change provides more control over when the tests are run, conserving resources.